pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-rasterizer"
version = "1.0.0"
description = "A GUI application for rasterizing redacted PDF files"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Pillow>=9.0.0",
    "PyPDF2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
]

[project.scripts]
pdf-rasterizer = "main:main"

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["../assets/*"]
